+++ src/test/java/school/redrover/PipelineProjectTest.java
@@ -1,12 +1,16 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
+
+import java.time.Duration;
 
 public class PipelineProjectTest extends BaseTest {
 
@@ -24,6 +28,22 @@ public class PipelineProjectTest extends BaseTest {
                 "» This field cannot be empty, please enter a valid name");
         Assert.assertTrue(
                 getDriver().findElement(By.id("ok-button")).isDisplayed());
+    }
+    @Test
+    public void testCreateWithDuplicateName() {
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.id("name")).sendKeys(PIPELINE_NAME);
+        getDriver().findElement(By.xpath("//span[text()='Pipeline']")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        getWait5().until(ExpectedConditions.presenceOfElementLocated(By.name("Submit")));
+        getDriver().findElement(By.id("jenkins-head-icon")).click();
+        getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//a[@href='/view/all/newJob']"))).click();
+
+        getDriver().findElement(By.id("name")).sendKeys(PIPELINE_NAME);
+        getDriver().findElement(By.xpath("//span[text()='Pipeline']")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.xpath("//div[@id='itemname-invalid']")).getText(),
+                "» A job already exists with the name ‘%s’".formatted(PIPELINE_NAME));
     }
     @Test
     public void testCreate() {
